+++ lib/src/app/alarm_manager.dart
@@ -3,5 +3,5 @@
 /// Provides access to the system alarm services.
 ///
 /// See: https://developer.android.com/reference/android/app/AlarmManager
-abstract class AlarmManager {
+class AlarmManager {
 }