+++ lib/src/app/keyguard_manager.dart
@@ -3,5 +3,5 @@
 /// Can be used to lock and unlock the keyboard.
 ///
 /// See: https://developer.android.com/reference/android/app/KeyguardManager
-abstract class KeyguardManager {
+class KeyguardManager {
 }