+++ lib/src/app/search_manager.dart
@@ -3,5 +3,5 @@
 /// Provides access to the system search services.
 ///
 /// See: https://developer.android.com/reference/android/app/SearchManager
-abstract class SearchManager {
+class SearchManager {
 }