+++ lib/src/os/battery_manager.dart
@@ -3,5 +3,5 @@
 /// This class provides a method for querying battery and charging properties.
 ///
 /// See: https://developer.android.com/reference/android/os/BatteryManager
-abstract class BatteryManager {
+class BatteryManager {
 }