+++ lib/src/os/power_manager.dart
@@ -3,5 +3,5 @@
 /// This class gives you control of the power state of the device.
 ///
 /// See: https://developer.android.com/reference/android/os/PowerManager
-abstract class PowerManager {
+class PowerManager {
 }