+++ lib/src/os/user_manager.dart
@@ -7,7 +7,7 @@ const Platform _platform = LocalPlatform();
 /// Manages users and user details on a multi-user system.
 ///
 /// See: https://developer.android.com/reference/android/os/UserManager
-abstract class UserManager {
+class UserManager {
 
   /// Returns whether this device supports multiple users with their own login
   /// and customizable space.