+++ test/android_app_test.dart
@@ -1,0 +1,12 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:flutter_android/android_app.dart';
+import 'package:flutter_android/android_content.dart' show Context;
+
+void main() {
+  group("android_app", () {
+    // TODO
+  });
+}