+++ test/android_test.dart
@@ -1,6 +1,6 @@
 /* This is free and unencumbered software released into the public domain. */
 
-import 'package:test/test.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 import 'package:flutter_android/android.dart';